[SOT] Move internal functions to a separate file
